refactor: move formatDateDisplay and formatLapTimeDisplay to dateUtils
