+++ app/(tabs)/addRide.tsx
@@ -23,7 +23,8 @@ import { useToggleLoading } from '@/hooks/useLoading'
 import useModalHandler from '@/hooks/useModalHandler'
 import { inputElementStyles } from '@/src/styles/InputStyles'
 import { AddableLap, AddableRide } from '@/src/types/AddableTypes'
-import { datetimeFieldToCapitals, formatDateForDisplay } from '@/src/utils/utils'
+import { formatDateForDisplay } from '@/src/utils/dateUtils'
+import { datetimeFieldToCapitals } from '@/src/utils/utils'
 import { router, useFocusEffect } from 'expo-router'
 import moment from 'moment-timezone'
 import React, { useEffect, useState } from 'react'
