+++ app/(tabs)/main/editRide.tsx
@@ -26,8 +26,8 @@ import useModalHandler from '@/hooks/useModalHandler'
 import { inputElementStyles } from '@/src/styles/InputStyles'
 import { AddableLap } from '@/src/types/AddableTypes'
 import { EditableRide } from '@/src/types/EditableTypes'
-import { getDateToIsoString } from '@/src/utils/dateUtils'
-import { datetimeFieldToCapitals, formatDateForDisplay } from '@/src/utils/utils'
+import { formatDateForDisplay, getDateToIsoString } from '@/src/utils/dateUtils'
+import { datetimeFieldToCapitals } from '@/src/utils/utils'
 import { router, useFocusEffect } from 'expo-router'
 import React, { useState } from 'react'
 import {
