+++ components/modal/addModal/AddLapModal.tsx
@@ -12,8 +12,7 @@ import useLocation from '@/hooks/useLocation'
 import useModalHandler from '@/hooks/useModalHandler'
 import { inputElementStyles } from '@/src/styles/InputStyles'
 import { AddableLap, AddableLapModalProp } from '@/src/types/AddableTypes'
-import { getDateToIsoString } from '@/src/utils/dateUtils'
-import { formatDateForDisplay, formatLapTimeDisplay } from '@/src/utils/utils'
+import { formatDateForDisplay, formatLapTimeDisplay, getDateToIsoString } from '@/src/utils/dateUtils'
 import { UserLocation } from '@maplibre/maplibre-react-native'
 import * as Crypto from 'expo-crypto'
 import { useFocusEffect } from 'expo-router'
