+++ components/modal/editModal/EditLapModal.tsx
@@ -11,8 +11,7 @@ import { useTheme } from '@/context/ThemeContext'
 import useModalHandler from '@/hooks/useModalHandler'
 import { inputElementStyles } from '@/src/styles/InputStyles'
 import { EditableLapModalProp } from '@/src/types/EditableTypes'
-import { getDateToIsoString } from '@/src/utils/dateUtils'
-import { formatDateForDisplay } from '@/src/utils/utils'
+import { formatDateForDisplay, getDateToIsoString } from '@/src/utils/dateUtils'
 import React, { useEffect, useState } from 'react'
 import { View } from 'react-native'
 import { ManageableLap } from '../FlatlistPicker'
