+++ components/ride/RideDetailCard.tsx
@@ -2,8 +2,7 @@ import { useTheme } from "@/context/ThemeContext"
 import { colors } from "@/src/const/color"
 import { travelDetailStyles } from "@/src/styles/TravelDetailStyles"
 import { CompleteRide } from "@/src/types/CompleteTypes"
-import { getDiffString } from "@/src/utils/dateUtils"
-import { formatLapTimeDisplay } from "@/src/utils/utils"
+import { formatLapTimeDisplay, getDiffString } from "@/src/utils/dateUtils"
 import moment from "moment"
 import { Text, View } from "react-native"
 import Container from "../Container"
